// src/app/interfaces/buttontable.interface.ts
export interface ButtonTableI {
    labelUpdate?: string;
    labelDelete?: string;
    labelAditional?: string;
    styleButtonUpdate?: string;
    styleButtonDelete?: string;
    styleButtonAditional?: string;
    iconButtonUpdate?: string;
    iconButtonDelete?: string;
    iconButtonAditional?: string;
    showUpdate?: boolean;
    showDelete?: boolean;
    showOption?: boolean;
}